rename uopgraph to rewriter [pr] (#8682)
