+++ test/external/external_benchmark_schedule.py
@@ -6,7 +6,7 @@ from tinygrad.ops import Ops
 from tinygrad.codegen.kernel import Kernel
 from tinygrad.codegen.lowerer import rewrite_shapetracker_with_index
 from tinygrad.codegen.linearize import linearize_uop
-from tinygrad.codegen.uopgraph import full_graph_rewrite
+from tinygrad.codegen.rewriter import full_graph_rewrite
 from tinygrad.engine.search import beam_search, bufs_from_lin
 
 if __name__ == "__main__":
