+++ test/test_renderer_failures.py
@@ -1,7 +1,7 @@
 import unittest
 from typing import List, cast
 import numpy as np
-from tinygrad.codegen.uopgraph import full_graph_rewrite
+from tinygrad.codegen.rewriter import full_graph_rewrite
 from tinygrad.codegen.linearize import linearize_uop
 from tinygrad.device import Buffer, Device
 from tinygrad.dtype import dtypes
