+++ test/unit/test_shapetracker.py
@@ -6,7 +6,7 @@ from tinygrad.helpers import prod
 from tinygrad.shape.shapetracker import ShapeTracker, View
 from tinygrad import Variable
 from tinygrad.ops import UOp, Ops, graph_rewrite
-from tinygrad.codegen.uopgraph import sym
+from tinygrad.codegen.rewriter import sym
 from itertools import product
 
 def shapetracker_getitem(st:ShapeTracker, val:int):
