+++ test/unit/test_simplify_valid_idx.py
@@ -1,6 +1,6 @@
 import unittest, itertools
 
-from tinygrad.codegen.uopgraph import full_graph_rewrite, is_increasing
+from tinygrad.codegen.rewriter import full_graph_rewrite, is_increasing
 from tinygrad.dtype import dtypes
 from tinygrad.ops import UOp, Ops, simplify_valid
