+++ tinygrad/codegen/rewriter.py
